+++ const_format/src/fmt/str_writer.rs
@@ -455,8 +455,6 @@ impl StrWriter {
     }
 }
 
-#[cfg_attr(feature = "docsrs", doc(cfg(feature = "const_generics")))]
-#[cfg(feature = "const_generics")]
 impl<const N: usize> StrWriter<[u8; N]> {
     /// Casts a `&StrWriter<[u8; N]>` to a `&StrWriter<[u8]>`,
     /// for calling methods defined on `StrWriter<[u8]>` (most of them).