+++ const_format/src/lib.rs
@@ -257,13 +257,9 @@
 //! allowing the `as_bytes_alt` methods and `slice_up_to_len_alt` methods to run
 //! in constant time, rather than linear time proportional to the truncated part of the slice.
 //!
-//! "const_generics":
-//! Enables impls that use const generics, currently only used for ergonomics.
-//! Use this when const generics are usable in stable Rust.
-//!
-//! "nightly_const_generics":
-//! Enables impls that use const generics, currently only used for ergonomics.
-//! This requires a nightly Rust compiler.
+//! - "const_generics": Requires Rust 1.51.0.
+//! Uses const generics in the implementation of the [`concatcp`] and [`formatcp`]
+//! macros to output less code.
 //!
 //!
 //!
@@ -353,6 +349,9 @@ pub mod panicking;
 
 mod pargument;
 
+#[cfg(feature = "const_generics")]
+mod const_generic_concatcp;
+
 #[cfg_attr(feature = "docsrs", doc(cfg(feature = "fmt")))]
 #[cfg(feature = "fmt")]
 pub mod utils;
@@ -420,6 +419,9 @@ pub mod pmr {
         option::Option::{self, None, Some},
         result::Result::{self, Err, Ok},
     };
+
+    #[cfg(feature = "const_generics")]
+    pub use crate::const_generic_concatcp::__priv_concatenate;
 
     #[cfg(feature = "fmt")]
     pub use crate::{